Fixed warnings in DevCpp, fixed join-lobby crash, added to_string and ftoa, download/upload calculations
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@897 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ include/Unicode.h
@@ -99,7 +99,7 @@ inline size_t Utf8StringSize(const std::string& str)  {
 	return res;
 }
 
-inline std::string Utf8SubStr(const std::string& str, size_t start, size_t n = -1) {
+inline std::string Utf8SubStr(const std::string& str, size_t start, size_t n = (size_t)-1) {
 	if (n == (size_t)-1)
 		return std::string(Utf8PositionToIterator(str, start), str.end());
 	else
@@ -108,7 +108,7 @@ inline std::string Utf8SubStr(const std::string& str, size_t start, size_t n = -
 			Utf8PositionToIterator(str, start + n));
 }
 
-inline void Utf8Erase(std::string& str, size_t start, size_t n = -1) {
+inline void Utf8Erase(std::string& str, size_t start, size_t n = (size_t)-1) {
 	std::string::iterator it = Utf8PositionToIterator(str, start);
 	str.erase(it, GetMultIncUtf8StringIterator(it, str.end(), n));	
 }
